+++ CHANGELOG.md
@@ -50,8 +50,9 @@
     * ... future intention is to make commands: `xl2json - RDINV`, `json2xml - WRXML`, `json2pdf`, `xml2roefact - LDXML` and other commands enumerated on `https://apitoroefact.renware.eu/commercial_agreement/110-SRE-api_to_roefact_requirements.html#componenta-xl2roefact`
 
 
-* wip... last used 231201piu_b
+* wip...
 
+* 231202piu_b `xl2roefact.py` started a skeleton for `file_name` argument - see code-in-file @"TODO here to use `excel_files_directory` + / + `file_name` to find all files and process them in a loop"
 * 231202piu_a build complete `cxFreeze` configuration in order to build Windows executable and installer package (as `msi` package):
     * [x] create a minimal setup (`setup.py`)
     * [x] create `build/` directory with all building commands for _xl2roefact_ app: `python setup.py build` (see official doc here `https://cx-freeze.readthedocs.io/en/latest/setup_script.html`)